refactored code for maintenace
