+++ server/controllers/incidents.controller.js
@@ -1,4 +1,4 @@
-import db from '../db/user.db';
+import db from '../db/db';
 import {
   getAllIncidents,
   createRecord,
